+++ src/filters/parser/SSFSplitter/SSFSplitter.cpp
@@ -204,7 +204,7 @@ HRESULT CSSFSplitterFilter::CreateOutputs(IAsyncReader* pAsyncReader)
 
 bool CSSFSplitterFilter::DemuxInit()
 {
-	SetThreadName(-1, "CSSFSplitterFilter");
+	SetThreadName((DWORD)-1, "CSSFSplitterFilter");
 	return true;
 }
